+++ lib/pay/braintree/billable.rb
@@ -37,8 +37,8 @@ module Pay
 
           result.customer
         end
-      rescue ::Braintree::AuthorizationError
-        raise Pay::Braintree::AuthorizationError
+      rescue ::Braintree::AuthorizationError => e
+        raise Pay::Braintree::AuthorizationError, e
       rescue ::Braintree::BraintreeError => e
         raise Pay::Braintree::Error, e
       end
@@ -57,8 +57,8 @@ module Pay
         raise Pay::Braintree::Error, result unless result.success?
 
         save_transaction(result.transaction)
-      rescue ::Braintree::AuthorizationError
-        raise Pay::Braintree::AuthorizationError
+      rescue ::Braintree::AuthorizationError => e
+        raise Pay::Braintree::AuthorizationError, e
       rescue ::Braintree::BraintreeError => e
         raise Pay::Braintree::Error, e
       end
@@ -84,8 +84,8 @@ module Pay
         raise Pay::Braintree::Error, result unless result.success?
 
         billable.create_pay_subscription(result.subscription, "braintree", name, plan, status: :active)
-      rescue ::Braintree::AuthorizationError
-        raise Pay::Braintree::AuthorizationError
+      rescue ::Braintree::AuthorizationError => e
+        raise Pay::Braintree::AuthorizationError, e
       rescue ::Braintree::BraintreeError => e
         raise Pay::Braintree::Error, e
       end
@@ -94,6 +94,8 @@ module Pay
       #
       # Returns true if successful
       def update_card(token)
+        customer unless processor_id?
+
         result = gateway.payment_method.create(
           customer_id: processor_id,
           payment_method_nonce: token,
@@ -107,8 +109,8 @@ module Pay
         update_card_on_file result.payment_method
         update_subscriptions_to_payment_method(result.payment_method.token)
         true
-      rescue ::Braintree::AuthorizationError
-        raise Pay::Braintree::AuthorizationError
+      rescue ::Braintree::AuthorizationError => e
+        raise Pay::Braintree::AuthorizationError, e
       rescue ::Braintree::BraintreeError => e
         raise Pay::Braintree::Error, e
       end